feat: enhance node key name (DATA-3508)
At several points in the node and node graph logic we map nodes to various
values. For the node key, we are using the name of the node which is the just
the name of the function which the node is declared with. However, this means
we'll get conflicts if we run a graph that has nodes with the same name.

The solution is to use a unique key for each node, the key is generated from
the qualified name of the node i.e the module path + class name + function
name which should be unique.