fix: repair bug causing an infinite loop in unit tests (DATA-3515)
We originally had a function pop_runnable_nodes that retrieved all runnable
nodes from the node graph and deleted them, I've tweaked this function to
only retrieve the nodes and delegated the removal of the nodes to the caller
of the function.

Unfortunately I didn't realise that this breaks if a node input is supplied
and thus becomes inactive, the get runnable nodes function will ignore it
when populating the list of runnable nodes but at the same time until it's
removed later nodes that depend upon it won't be runnable.

The fix is to remove all inactive nodes from the execution graph, I wanted to
do this on the node graph side as it pertains more to the node graph but it
doesn't make sense outside the context of an execution graph, therefore I
changed the copy function to get_execution_graph to make it a more rationale
home for a command to wipe all the inactive nodes.