fix: drag and drop new node doesn't work if examining a node with a tag
Whenever we have the node sidebar open with at least one tag, the drag and drop
doesn't work. This is due to a bug with react-tags
(https://github.com/react-tags/react-tags/issues/723), it ought to have been
fixed but still seems to be replicable in our code.

I've solved this by swapping out react-tags for a different library called React
Tag Autocomplete.